Move DatabaseOverview to pages
